assss

+++ backend/controllers/InfoPublicController.php
@@ -132,7 +132,9 @@ class InfoPublicController extends Controller
                     mkdir($tmp, 0777, true);
                 }
                 if ($image_header->saveAs($tmp . $file_name)) {
-                    unlink($tmp.$old_image_header);
+                    if(file_exists($tmp.$old_image_header)){
+                        unlink($tmp.$old_image_header);
+                    }
                     $model->image_header = $file_name;
                 }
             }else{
@@ -146,7 +148,9 @@ class InfoPublicController extends Controller
                     mkdir($tmp, 0777, true);
                 }
                 if ($image_footer->saveAs($tmp . $file_name)) {
-                    unlink($tmp.$old_image_footer);
+                    if(file_exists($tmp.$old_image_footer)){
+                        unlink($tmp.$old_image_footer);
+                    }
                     $model->image_footer = $file_name;
                 }
             }else{
@@ -160,7 +164,9 @@ class InfoPublicController extends Controller
                     mkdir($tmp, 0777, true);
                 }
                 if ($image_menu->saveAs($tmp . $file_name)) {
-                    unlink($tmp.$old_image_menu);
+                    if(file_exists($tmp.$old_image_menu)){
+                        unlink($tmp.$old_image_menu);
+                    }
                     $model->image_menu = $file_name;
                 }
             }else{
